fix(forest): properly handle createStore in classList
In forect, when creating the css class store with createStore directly — without additional call to restore — the initial value was not being applied correctly to the classList.

The fix creates a new store with the correct initial value to handle updates, ensuring the initial state is properly set and all the further updates to the class store is being applied.

Fixes #965